Change ints to longs in PageInfo. We are writing longs into it in CFMod

+++ src/cf.nova.h
@@ -189,8 +189,8 @@ struct FileLine_
 
 typedef struct
    {
-   int resultsPerPage;
-   int pageNum;
+   long resultsPerPage;
+   long pageNum;
    int totalResultCount;
    } PageInfo;
